requirements.txt
descartes == 1.1.0
h5py == 2.8.0
numpy == 1.15.1
matplotlib == 3.0.0
pandas == 0.23.4
pyexcel == 0.5.9.1
pyexcel-xls == 0.5.8
pyexcel-xlsx == 0.5.6
pyshp == 1.2.12
Shapely == 1.6.4.post2
scikit-learn == 0.20.0
scipy == 1.1.0
python-vlc == 3.0.7110
mutagen == 1.42.0
